requirements.txt
python-telegram-bot==21.0.1
rich
fuzzywuzzy
python-Levenshtein
python-telegram-bot[job-queue]

# Requirements for DeepDanbooru
git+https://github.com/KichangKim/DeepDanbooru.git
Click>=7.0
numpy>=1.16.2
scikit-image>=0.15.0
tensorflow>=2.7.0
tensorflow-io>=0.22.0
requests>=2.22.0
six>=1.13.0
Pillow~=9.4.0